Dihedral Activation Functions
Made the dihedral activation function class, with 3 types
- S: if the rotation causes a reflection, it is changed to the identity
- R: if the rotation rotates a shape more than pi (s invariant), it is changed to the identity
- B: if the rotation puts a vertex into a position where it would be fastest to get there through a clockwise rotation, reflection invariant, it is changed to the identity

These were made as one class, with a type variable, rather than the activation function layout, which is the same under the hood, but in Python shows a series of classes. I chose to do this because it seemed simpler, and may eventually split them.

S is for the dihedral reflection s, R is for dihedral rotation r, B is for bottom